Split up the code with h file as example.
